calorie-counter-code/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Keras_Preprocessing == 1.1.2
Pillow == 9.1.1
boto3 == 1.24.5
fuzzywuzzy == 0.18.0
gensim == 4.2.0
numpy == 1.22.3
psutil == 5.9.1
python_Levenshtein == 0.12.2
sklearn == 0.0
torch == 1.11.0
transformers == 4.19.2
